feat(ppsci): remove redundant codes
